CMakeLists.txt, small changes to make building for clang and gcc easier

+++ nina-chess/psqt.h
@@ -17,7 +17,7 @@ public:
 	forceinline constexpr void Reset(const Position& position, const MoveList& moveList);
 	forceinline constexpr void IncrementalUpdate(const Position& position, const MoveList& moveList);
 	forceinline constexpr void UndoUpdate() { m_Depth--; }
-	forceinline float Evaluate() { return std::tanhf(m_Accumulators[m_Depth].GetOutput()[0] / 16.f); }
+	forceinline float Evaluate() { return std::tanh(m_Accumulators[m_Depth].GetOutput()[0] / 16.f); }
 
 private:
 	forceinline constexpr void update(const Position& position, const MoveList& moveList);
